Add files via upload
It’s my first time doing that and I have no idea if it worked. The game is pretty simple, use the arrow keys to move the black cat and avoid the other ones. One of them will follow the black cat while the other will move randomly. If the random cat touches the black one, it (black cat) will grow, while if the following cat touches the black cat, it (following cat) will move to a random location and continue following the black cat. I think I covered it all.